+++ bot/events/room_playlist-queue-update-dub.js
@@ -6,6 +6,7 @@
 
 'use strict';
 const historyStore = require(process.cwd()+ '/bot/store/history.js');
+const queue = require(process.cwd()+ '/bot/utilities/dj.js');
 const _ = require('lodash');
 const repo = require(process.cwd()+'/repo');
 
@@ -70,11 +71,43 @@ module.exports = function(bot, db) {
       return bot.log('error', 'BOT', 'roomPlaylistQueueUpdate: data object missing');
     }
 
-    if (Array.isArray(data.queue) && data.queue.length > 0) {
+    if (!Array.isArray(data.queue)) {
+      return bot.log('error', 'BOT', 'queue is not an array');
+    }
+
+    if (data.queue.length > 0) {
       data.queue.forEach(function(q){
         checkHistory(bot, q);
         checkNewUser(bot, db, q.user);
       });
     }
+
+    // join the queue when no one is in it
+    if (data.queue.length === 0) {
+      bot.sendChat("Looks like no one is in the queue, guess I'll take over for a while");
+      queue.join(bot, function(code,_data, extra){
+        if (code === 200) {
+          bot.log('info', 'BOT', 'Successfully joined the queue when no one was in it');
+          bot.isDJing = true;
+          bot.commandedToDJ = false;
+        } else {
+          bot.log('error','BOT', 'Could not un-pause my queue');
+        }
+      });
+      return;
+    }
+
+    // leave the queue when when a new person joins it
+    if (data.queue.length > 1 && bot.isDJing && !bot.commandedToDJ) {
+      bot.sendChat("Leaving the queue now that other DJs are in it");
+      queue.leave(bot, function(code, _data, extra){
+        if (code === 200) {
+          bot.log('info', 'BOT', 'Successfully cleared my queue');
+        } else {
+          bot.log('error','BOT', `Could not clear the queue - ${code} ${_data}`);
+        }
+      });
+    }
+
   });
 };