+++ bot/index.js
@@ -14,6 +14,7 @@ var db = new Database(svcAcct, BASEURL);
 DubAPI.prototype.getRoomHistory = require(process.cwd() + '/bot/extend/getRoomHistory.js');
 DubAPI.prototype.addToPlaylist = require(process.cwd() + '/bot/extend/addToPlaylist.js');
 DubAPI.prototype.getPlaylists = require(process.cwd() + '/bot/extend/getPlaylists.js');
+DubAPI.prototype.shufflePlaylist = require(process.cwd() + '/bot/extend/shufflePlaylist.js');
 
 new DubAPI({ username: settings.USERNAME, password: settings.PASSWORD }, function(err, bot) {
         
@@ -25,6 +26,8 @@ new DubAPI({ username: settings.USERNAME, password: settings.PASSWORD }, functio
     
     bot.myconfig = config;
     bot.maxChatMessageSplits = 5;
+    bot.commandedToDJ = false;
+    bot.isDJing = false;
 
     if (bot.myconfig.verboseLogging) {
       bot.log = require('jethro');